pkg/asset/internal: Introduce a FLock around the API server
This commit represents a workaround for #262. By maintaining
a file lock while the API server is running (either temporary
or self-hosted), we prevent the self-hosted API server from
starting and trying to bind ports, until the temporary one is
stopped. Therefore, we avoid the loop where the self-hosted
API server would crash as soon as it is brought up due to the
ports already being bound by the stopping temporary server.