templates: ensure critical components dont coexist
This commit adds antiaffinity rules to the controller manager and the
sheduler to ensure that these deployments' pods do not all end up on the
same master node, especially now that they are pinned with node
selectors. To make this possible on small clusters, e.g. those with 1
master node, this is specified with "preferred" rather than "required".