Preliminary recovery command for bootkube. (#491)
The new `bootkube recover` command provides a framework for reading
control plane manifests out of a storage backend, serializing them to
disk, and constructing pod manifests for the bootstrap components that
are usable with `bootkube start` to re-bootstrap the control plane.

The command initially provides and etcdclient backend for use with a
running etcd cluster. In the future we can add support other methods,
such as an etcd backup directory or a running apiserver.

There is also documentation and an example script in hack/multi-node
that deletes a master node, recovers the assets from etcd, and then
re-bootstraps the control plane.