hack/terraform-quickstart: remove dependancy on default network
This PR avoids terraform depending on a sane default network existing to
expose the nodes. This includes creating a VPC, subnet, route table,
internet gateway, open security group, and a network ACL.  This will be
useful once tools to auto-wipe a region are used so that it doesn't take
out all our tests that depend on default networks lying around.