checkpointer: add test case asserting behavior.
The checkpointer will activate a checkpoint if the local kubelet API
parent exists but the CRI shim doesn't report it running. We may want
to change this later, but this is the current behavior.

Also make the stop and remove orders stable to make the tests
consistently pass and also because why not.