e2e: make checkpointer tests more robust.
Instead of extending the existing checkpointer daemonset to run on
worker nodes, start a new checkpointer daemonset instead. This makes
sure that the existing checkpointer (on the apiserver) is not disturbed.

Furthermore, this allows the checkpointer tests to work with the new
checkpointer that is only allowed to look at its own namespace.