Makefile: Use 'git diff --exit-code' for the gofmt check
This makes fixing errors easier, because you can just commit/amend
after a failed 'make gofmt'.  The diff check failes the target if
there are any changes (either from gofmt or otherwise), which is good
for continuous integration testing and such.

Splitting gofmt out into its own .PHONY rule makes it more obvious
that the diff call is part of the check, and allows you to run the
gofmt check without pulling in the other, unrelated 'check' tests.

This also reduces our shell exposure.  The read and [ bits looked ok
to me, but this was the only place we called read in the Makefile, so
I expect it was the source of [1]:

  05:22:18 [1650506.967906] golang[6]: /bin/sh: 1: read: arg count

which was part of the motivation for 132a3815 (use bash for the
makefile shell, 2018-01-29, #870).

[1]: https://github.com/kubernetes-incubator/bootkube/pull/870#issue-165889403