wait for self hosted control plane to be available
bootstrap API server should terminate only after API is HA
bootstrap API server should be terminated only after API is
HA, which means we wait for API to be available on at
least 2 master nodes.

- a) we achieve HA API
- b) delete the bootstrap kube-apiserver manifests
- c) wait for the bootstrap API to de down
- d) delete all other static manifests
- e) mark the bootstrap process done

We should note the difference between the bootstrap node itself,
and API being available on the bootstrap node.
Today inside the cluster bootstrap, we remove the bootstrap API
as soon as two master nodes appear. So this is what happens
today on the bootstrap node:

a) create the static assets
b) wait for 2 master nodes to appear
c) remove the kube-apiserver from the bootstrap node
d) mark the bootstrap process as completed
e) the bootkube script waits for api to be HA

But we already might have a window of API not available starting
from c and until api is available on a master node.

This commit waits until API is HA (API is available on at least
two master nodes) before we tear down the bootstrap control plane

We also wait for scheduler, and kcm to be available on at least
two master nodes